requirements.txt
selenium
pytest
pyyaml
webdriver-manager
allure-pytest